Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4707



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@18973 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/ucol_tok.cpp
@@ -1,7 +1,7 @@
 /*
 *******************************************************************************
 *
-*   Copyright (C) 2001-2005, International Business Machines
+*   Copyright (C) 2001-2006, International Business Machines
 *   Corporation and others.  All Rights Reserved.
 *
 *******************************************************************************
@@ -1554,7 +1554,7 @@ uint32_t ucol_tok_assembleTokenList(UColTokenParser *src, UParseError *parseErro
           } else { /* there is both explicit and implicit expansion. We need to make a combination */
             uprv_memcpy(src->extraCurrent, src->source + (expandNext & 0xFFFFFF), (expandNext >> 24)*sizeof(UChar));
             uprv_memcpy(src->extraCurrent+(expandNext >> 24), src->source + src->parsedToken.extensionOffset, src->parsedToken.extensionLen*sizeof(UChar));
-            sourceToken->expansion = (uint32_t)(((expandNext >> 24) + src->parsedToken.extensionLen)<<24 | (src->extraCurrent - src->source));
+            sourceToken->expansion = (uint32_t)(((expandNext >> 24) + src->parsedToken.extensionLen)<<24 | (uint32_t)(src->extraCurrent - src->source));
             src->extraCurrent += (expandNext >> 24) + src->parsedToken.extensionLen;
           }
         }
